src/main/webapp/app/entities/danh-muc-tinh/delete/danh-muc-tinh-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucTinhService } from '../service/danh-muc-tinh.service';

import { DanhMucTinhDeleteDialogComponent } from './danh-muc-tinh-delete-dialog.component';

describe('DanhMucTinh Management Delete Component', () => {
  let comp: DanhMucTinhDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucTinhDeleteDialogComponent>;
  let service: DanhMucTinhService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucTinhDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucTinhDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucTinhDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucTinhService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
